Introduce easyfile common module (#2)
