+++ default.nix
@@ -8,5 +8,6 @@ with import <nixpkgs> {}; {
         ./venv/bin/pip install -r requirements.txt
       fi
     '';
+    PYTHONPATH = "./commons/";
   };
 }
